Refactor player exit hook
